requirements.txt
Django==3.1.3
python-dateutil==2.8.1
djangorestframework==3.12.2
unittest-xml-reporting==3.0.4
requests==2.25.0
pytest==6.2.5
pytest-django==4.5.2
django-filter==2.4.0
